completed exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -18,7 +18,7 @@ blogsRouter.post('/', async (request, response) => {
   
   // decode the JWT
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if(!decodedToken.id) {
+  if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
 
@@ -37,6 +37,16 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blogToDelete = await Blog.findById(request.params.id)
+  if (!blogToDelete) {
+    return response.status(204).end()
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id || decodedToken.id.toString() !== blogToDelete.user.toString()) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
